// Server/apiplate/ClientApp/src/app/shared/form-builder/models/control-type.enum.ts
export enum ControlTypes {
  TextInput = "text",
  PasswordInput = "password",
  NumberInput = "number",
  Selection = "selection",
  MultiSelection = "multi-selection",
  CheckBox = "checkbox",
  DatePicker = "datepicker",
  LocalFilePicker = "LocalFilePicker",
  Url = "url",
  Hidden = "hidden",
  TextArea = "TextArea",
  RichTextEditor = "RichTextEditor",
  MapPicker = "MapPicker",
  TableBuilder = "TableBuilder"
}